requirements.txt
mysql-connector-python==8.2.0
Jinja2==3.1.2
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.7
Pillow==10.1.0
bcrypt==4.1.2